feat: update bk repo api
